backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.1
pydantic-settings==2.1.0
motor==3.3.1
pymongo==4.6.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
scikit-learn==1.3.2
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
openai==1.3.8
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
email-validator==2.1.0
jinja2==3.1.2
itsdangerous==2.1.2
bcrypt==4.0.1
cryptography==41.0.5
requests==2.31.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
neurokit2==0.2.6
scipy==1.11.4
langchain==0.0.342
langchain-openai==0.0.2
chromadb==0.4.18
beautifulsoup4==4.12.2
PyPDF2==3.0.1
tiktoken==0.5.1 